+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.0 (WIP)
 
 * Make sure `String` methods don't return object versions of the string
+* Fix `Pledge.done()` resolving with object versions of primitives
 
 ## 0.8.18 (2024-01-19)