+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `Blast.parseClassPath(path)` method
 * Add `WeakValueSet` class
 * Add `WeakValueSetMap` class
+* Make `Blast.createObjectId()` also available on the server-side
 
 ## 0.8.18 (2024-01-19)