+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Add `WeakValueSetMap` class
 * Make `Blast.createObjectId()` also available on the server-side
 * Add `toISOString` method to `AbstractDateTime` classes
+* Add `Blast.didInitialLoad()` method
+* Make `Blast.executeAfterLoadingCycle()` keep `did_initial_load` variable alone
 
 ## 0.8.18 (2024-01-19)