+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Make `Blast.executeAfterLoadingCycle()` keep `did_initial_load` variable alone
 * Add `String.parseQuoted(input)` to parse json-like string values
 * Add the `Placeholder` and `Trail` class to the Develry namespace
+* The `Pledge` class is now a child of `Placeholder` and no longer an `Informer`
 
 ## 0.8.18 (2024-01-19)