+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Add new operators to the JavaScript tokenizer
 * Replace `Object.hasOwnProperty.call` calls with `Object.hasOwn`
 * Add support for underscores in numbers to the tokenizer
+* Remove some unused string methods
 
 ## 0.8.18 (2024-01-19)