+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Replace `Object.hasOwnProperty.call` calls with `Object.hasOwn`
 * Add support for underscores in numbers to the tokenizer
 * Remove some unused string methods
+* Removed special Bun require code
 
 ## 0.8.18 (2024-01-19)